remove function returns
