+++ src/index.ts
@@ -1,25 +1,20 @@
 export {
   Playthrough,
   usePlaythrough,
+  usePlaythroughContext,
   type PlaythroughProps,
 } from "./Playthrough.js";
-export {
-  PlaythroughConsumer,
-  PlaythroughContext,
-  type PlaythroughContextType,
-  PlaythroughProvider,
-  usePlaythroughContext,
-} from "./PlaythroughContext.js";
 
 export {
   Conversation,
   useConversation,
+  ChatMode,
+  useConversationContext,
   type UseConversationOptions,
   type ConversationProps,
   type ConversationChildProps,
   type ConversationState,
   type Conversation as ConversationType,
-  ChatMode,
   type StoredMessage,
   type PlayerMessage,
 } from "./Conversation.js";
@@ -37,7 +32,6 @@ export {
   type QueuedConversationProps,
   type QueuedConversationChildProps,
 } from "./QueuedConversation.js";
-export { useConversationContext } from "./ConversationContext.js";
 
 export {
   default as useBackgroundVideo,
